+++ src/discrete-layer/models/LayerDemRecordModel.base.ts
@@ -102,7 +102,7 @@ export class LayerDemRecordModelSelector extends QueryBuilder {
   get insertDate() { return this.__attr(`insertDate`) }
   get wktGeometry() { return this.__attr(`wktGeometry`) }
   get keywords() { return this.__attr(`keywords`) }
-  links(builder?: string | LinkModelSelector | ((selector: LinkModelSelector) => LinkModelSelector)) { return this.__child(`links`, LinkModelSelector, builder) }
+  links(builder: string | LinkModelSelector | ((selector: LinkModelSelector) => LinkModelSelector) | undefined) { return this.__child(`links`, LinkModelSelector, builder) }
 }
 export function selectFromLayerDemRecord() {
   return new LayerDemRecordModelSelector()